tests/wpt/web-platform-tests/css/css-ui/parsing/resize-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing resize with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#resize">
<meta name="assert" content="resize supports the full grammar 'none | both | horizontal | vertical'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("resize", "none");
test_valid_value("resize", "both");
test_valid_value("resize", "horizontal");
test_valid_value("resize", "vertical");
</script>
</body>
</html>
